+++ resources/ReleaseNotes.md
@@ -1,3 +1,4 @@
+- Improve time estimates of mode change events
 - Hide unnecessary constants in meta data classes
 - Make sure custom events have padded y-axis limits
 - Fix issue with consecutive figures not opening maximized